+++ tests/regression_tests_data.py
@@ -14,7 +14,7 @@
 # File contributors : Étienne André, Jaime Arias
 #
 # Created           : 2015/10/23
-# Last modified     : 2021/02/02
+# Last modified     : 2021/02/05
 #************************************************************
 
 
@@ -372,6 +372,119 @@ Is strongly deterministic?              : false
 
 	,
 
+
+	#*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-
+	# TESTS ON UNSATISFIABLE INITIAL STATE
+	#*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-
+
+	#------------------------------------------------------------
+	{
+		# Test version             : 1
+		# Test since               : 2021/02/05
+		# Last modified            : 2021/02/05
+		# Test for IMITATOR version: 3
+		'purpose'    : 'Test unsatisfiable initial state (clock constraint)',
+		'input_files': ['unsatisfiableInitStateClocks.imi'],
+		'options'    : '-mode statespace',
+		'expectations' : [
+			{'file': 'unsatisfiableInitStateClocks.res' , 'content' : """
+Error                                   : unsatisfiable initial state
+		"""
+			} # end result file
+			,
+		] # end expectations
+	} # end test case
+	#------------------------------------------------------------
+
+	,
+	
+	#------------------------------------------------------------
+	{
+		# Test version             : 1
+		# Test since               : 2021/02/05
+		# Last modified            : 2021/02/05
+		# Test for IMITATOR version: 3
+		'purpose'    : 'Test unsatisfiable initial state (clock constraint + invariant)',
+		'input_files': ['unsatisfiableInitStateClocksInv.imi'],
+		'options'    : '-mode statespace',
+		'expectations' : [
+			{'file': 'unsatisfiableInitStateClocksInv.res' , 'content' : """
+Error                                   : unsatisfiable initial state
+		"""
+			} # end result file
+			,
+		] # end expectations
+	} # end test case
+	#------------------------------------------------------------
+
+	,
+	
+	#------------------------------------------------------------
+	{
+		# Test version             : 1
+		# Test since               : 2021/02/05
+		# Last modified            : 2021/02/05
+		# Test for IMITATOR version: 3
+		'purpose'    : 'Test unsatisfiable initial state (discrete+clock constraint)',
+		'input_files': ['unsatisfiableInitStateDiscrete.imi'],
+		'options'    : '-mode statespace',
+		'expectations' : [
+			{'file': 'unsatisfiableInitStateDiscrete.res' , 'content' : """
+Error                                   : unsatisfiable initial state
+		"""
+			} # end result file
+			,
+		] # end expectations
+	} # end test case
+	#------------------------------------------------------------
+
+	,
+	
+	#------------------------------------------------------------
+	{
+		# Test version             : 1
+		# Test since               : 2021/02/05
+		# Last modified            : 2021/02/05
+		# Test for IMITATOR version: 3
+		'purpose'    : 'Test unsatisfiable initial state (incompatible discrete invariant)',
+		'input_files': ['unsatisfiableInitStateDiscreteInv.imi'],
+		'options'    : '-mode statespace',
+		'expectations' : [
+			{'file': 'unsatisfiableInitStateDiscreteInv.res' , 'content' : """
+Error                                   : unsatisfiable initial state
+		"""
+			} # end result file
+			,
+		] # end expectations
+	} # end test case
+	#------------------------------------------------------------
+
+	,
+	
+	#*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-
+	# TESTS ON DISCRETE VARIABLES
+	#*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-
+
+	#------------------------------------------------------------
+	{
+		# Test version             : 1
+		# Test since               : 2021/02/05
+		# Last modified            : 2021/02/05
+		# Test for IMITATOR version: 3
+		'purpose'    : 'Test division by 0',
+		'input_files': ['divisionby0.imi'],
+		'options'    : '-mode statespace',
+		'expectations' : [
+			{'file': 'divisionby0.res' , 'content' : """
+Error                                   : division by 0
+		"""
+			} # end result file
+			,
+		] # end expectations
+	} # end test case
+	#------------------------------------------------------------
+
+	,
 
 	##------------------------------------------------------------
 	{