+++ RELEASES.md
@@ -5,7 +5,8 @@ minor release
 ### Result
 * IMITATOR now attempts to generate a result (`.res`) file even in the case of an error such as parsing error of the model or of the property
 
-
+### Bug fixing:
+* Fixed an issue met in v3 when a division by 0 is encountered.
 
 
 *******************************************************